requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.7
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==2.0.0
idna==3.3
Pillow==8.3.2
pytz==2022.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
